+++ app/models/patient.rb
@@ -17,6 +17,7 @@ class Patient < ApplicationRecord
 
   has_many :data_files, dependent: :destroy
   has_many :computations, dependent: :destroy
+  has_many :pipelines, dependent: :destroy
 
   validates :case_number, :procedure_status, presence: true
   validates :case_number, uniqueness: true